refactor: extract input.rs
